Group all module containers in their own folder
